+++ src/Celtchar/Novel/Structure.hs
@@ -48,17 +48,22 @@ instance Show Language where
     show English = "english"
     show French  = "french"
 
-data Novel = Novel { author     :: String
-                   , language   :: Language
-                   , novelTitle :: String
-                   , manuscript :: Manuscript }
+data Novel = Novel { author      :: String
+                   , language    :: Language
+                   , novelTitle  :: String
+                   , frontmatter :: Maybe [Chapter]
+                   , manuscript  :: Manuscript
+                   , appendix    :: Maybe [Chapter]
+                   }
   deriving (Generic, Show)
 
 instance FromJSON Novel where
     parseJSON (Object v) = Novel <$> v .: "author"
                                  <*> v .: "language"
                                  <*> v .: "title"
+                                 <*> v .: "frontmatter"
                                  <*> v .: "manuscript"
+                                 <*> v .: "appendix"
 
 getNovelStructure :: FilePath -> IO (Either String Novel)
 getNovelStructure conf = do