+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## celtchar-0.2.0.0
 
-### Feature
+### Features
 
 * If `celtchar` cannot parse the configuration file, it reports an error
+* The novel can have a frontmatter and an appendix
 
 ## celtchar-0.1.0.1